+++ lib/jekyll/polyglot/liquid.rb
@@ -3,6 +3,7 @@ module Jekyll
     module Liquid
       require_relative 'liquid/tags/i18n_headers'
       require_relative 'liquid/tags/static_href'
+      require_relative 'liquid/filters/lang_slug'
     end
   end
 end